Rename `thprac_utils.*` to `common.*`
