+++ thprac/src/thprac/common.cpp
@@ -1,4 +1,4 @@
-﻿#include "thprac_utils.h"
+﻿#include "common.h"
 #include "thprac_licence.h"
 #include "thprac_launcher_main.h"
 #include "thprac_launcher_cfg.h"
