+++ thprac/src/thprac/common.h
