+++ thprac/src/thprac/thprac_alcostg.cpp
@@ -1,4 +1,4 @@
-﻿#include "thprac_utils.h"
+﻿#include "common.h"
 
 
 namespace THPrac {
