+++ thprac/src/thprac/thprac_hook.cpp
@@ -3,7 +3,7 @@
 #define NOMINMAX
 
 #include "thprac_launcher_cfg.h"
-#include "thprac_utils.h"
+#include "common.h"
 #include "thprac_gui_locale.h"
 
 #include "..\MinHook\src\buffer.h"
