+++ thprac/src/thprac/thprac_launcher_wnd.cpp
@@ -4,7 +4,7 @@
 #include "thprac_gui_impl_win32.h"
 #include "thprac_gui_locale.h"
 #include "thprac_launcher_utils.h"
-#include "thprac_utils.h"
+#include "common.h"
 #include "..\..\resource.h"
 #include <d3d9.h>
 #include <tchar.h>
