+++ thprac/src/thprac/thprac_th095.cpp
@@ -1,4 +1,4 @@
-﻿#include "thprac_utils.h"
+﻿#include "common.h"
 
 
 namespace THPrac {
