+++ thprac/thprac.vcxproj
@@ -69,7 +69,7 @@
     <ClInclude Include="src\thprac\thprac_gui_locale.h" />
     <ClInclude Include="src\thprac\thprac_locale_def.h" />
     <ClInclude Include="src\thprac\attach.h" />
-    <ClInclude Include="src\thprac\thprac_utils.h" />
+    <ClInclude Include="src\thprac\common.h" />
     <ClInclude Include="src\thprac\thprac_version.h" />
     <ClInclude Include="src\thprac\utils\utils.h" />
   </ItemGroup>
@@ -135,7 +135,7 @@
     <ClCompile Include="src\thprac\thprac_th17.cpp" />
     <ClCompile Include="src\thprac\thprac_th18.cpp" />
     <ClCompile Include="src\thprac\thprac_th185.cpp" />
-    <ClCompile Include="src\thprac\thprac_utils.cpp" />
+    <ClCompile Include="src\thprac\common.cpp" />
     <ClCompile Include="src\thprac\thprac_version.cpp" />
   </ItemGroup>
   <ItemGroup>
